Refactor tenant services and add utility functions
This update includes a refactoring of the tenant services within the PDA application. The TenantRepository was switched to TenantsRepository in methods to create a tenant and get a tenant by id. The code related to tenant creation was also compressed for clarity. Additionally, new utility functions were added to convert between Unix time and datetime objects. Some refactoring was conducted on properties and other infrastructure components too.